requirements.txt
torch
ultralytics
kafka-python
Pillow
